requirements.txt
aiogram==3.3.0
Pillow==10.1.0
python-dotenv==1.0.0
APScheduler==3.10.4
environs==10.3.0
pytz==2023.3 